Add vue
